refactor(schedule): centralize Turno type
