+++ src/api/schedule.ts
@@ -1,20 +1,8 @@
 import api from './axios'
+import type { Turno } from 'src/types/turno'
 
-export interface Turno {
-  id: string
-  user_id: string
-  slot1: { inizio: string; fine: string }
-  slot2?: { inizio: string; fine: string } | null
-  slot3?: { inizio: string; fine: string } | null
-}
-
-export const listTurni = (): Promise<Turno[]> =>
-  api.get<Turno[]>('/orari/').then(r => r.data)
-
-export const createTurno = (
-  data: Omit<Turno, 'id'>
-): Promise<Turno> =>
-  api.post<Turno>('/orari/', data).then(r => r.data)
+export const fetchTurni = () => api.get<Turno[]>('/orari/')
+export const saveTurno = (t: Turno) => api.post<Turno>('/orari/', t)
 
 export const deleteTurno = (id: string): Promise<void> =>
   api.delete(`/orari/${id}`).then(() => undefined)
