+++ src/pages/SchedulePage.tsx
@@ -1,5 +1,7 @@
 import React, { useEffect, useState } from 'react';
 import api from '../api/axios';
+import { fetchTurni, saveTurno } from '../api/schedule';
+import type { Turno } from '../types/turno';
 import { listUtenti, Utente } from '../api/users';
 import { getSchedulePdf } from '../api/pdfs';
 import { format, startOfISOWeek, addDays } from 'date-fns';
@@ -10,16 +12,6 @@ import './ListPages.css';
 
 /* ---------- TIPI ---------- */
 interface Slot { inizio: string; fine: string; }
-interface Turno {
-  id: string;
-  giorno: string;
-  slot1: Slot;
-  slot2?: Slot;
-  slot3?: Slot;
-  tipo: 'NORMALE' | 'STRAORD' | 'FERIE' | 'RIPOSO' | 'FESTIVO';
-  note?: string;
-  user_id: string;
-}
 
 interface NewTurnoPayload {
   user_id: string;
@@ -64,9 +56,9 @@ export default function SchedulePage() {
   const [signedIn, setSignedIn] = useState(false);
   const [signInError, setSignInError] = useState('');
 
-  const fetchTurni = async () => {
+  const loadTurni = async () => {
     try {
-      const { data } = await api.get<Turno[]>('/orari/');
+      const { data } = await fetchTurni();
       setTurni(data);
       setLoadError('');
       return data;
@@ -78,7 +70,7 @@ export default function SchedulePage() {
 
   const handleImportComplete = async (success: boolean) => {
     if (success) {
-      const data = await fetchTurni();
+      const data = await loadTurni();
       if (data.length) {
         const maxDate = new Date(
           Math.max(...data.map(t => new Date(t.giorno).getTime()))
@@ -118,7 +110,7 @@ export default function SchedulePage() {
       setUtenti(r.data);
       setUtenteSel(r.data[0]?.id ?? '');
     });
-    void fetchTurni();
+    void loadTurni();
   }, []);
 
   useEffect(() => {
@@ -158,7 +150,7 @@ export default function SchedulePage() {
     if (s2Start && s2End) payload.slot2 = { inizio: s2Start, fine: s2End };
     if (s3Start && s3End) payload.slot3 = { inizio: s3Start, fine: s3End };
 
-    const { data } = await api.post<Turno>('/orari/', payload);
+    const { data } = await saveTurno(payload as Turno);
     setTurni(prev =>
       prev.some(t => t.id === data.id) ? prev.map(t => t.id === data.id ? data : t) : [...prev, data]
     );
